+++ pytket/docs/changelog.rst
@@ -14,6 +14,8 @@ Minor new features:
 * New ``CustomPass()`` accepting a user-supplied circuit transformation
   function.
 * ``measure_register`` now allows using an existing classical register
+* Provide an additional ``RebaseCustom`` constructor that takes a
+  TK2-replacement instead of a CX-replacement function.
 
 1.4.3 (July 2022)
 -----------------